--- budget-binder-multiplatform-app/src/commonMain/kotlin/de/hsfl/budgetBinder/presentation/viewmodel/LogoutViewModel.kt
@@ -1,36 +1,0 @@
-package de.hsfl.budgetBinder.presentation.viewmodel
-
-import de.hsfl.budgetBinder.common.DataResponse
-import de.hsfl.budgetBinder.domain.usecase.LogoutUseCase
-import de.hsfl.budgetBinder.presentation.UiState
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-
-class LogoutViewModel(
-    private val logoutUseCase: LogoutUseCase,
-    private val scope: CoroutineScope = CoroutineScope(Dispatchers.Unconfined + SupervisorJob())
-) {
-    private val _state = MutableStateFlow<UiState>(UiState.Empty)
-    val state: StateFlow<UiState> = _state
-
-    fun logOut(onAllDevices: Boolean) {
-        logoutUseCase(onAllDevices).onEach {
-            when (it) {
-                is DataResponse.Success -> {
-                    _state.value = UiState.Success(it.data)
-                }
-                is DataResponse.Error -> {
-                    _state.value = UiState.Error(error = it.message ?: "Something went wrong")
-                }
-                is DataResponse.Loading -> {
-                    _state.value = UiState.Loading
-                }
-            }
-        }.launchIn(scope)
-    }
-}
