+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/user/UserComponent.kt
@@ -1,14 +1,7 @@
 package de.hsfl.budgetBinder.compose.user
 
 import androidx.compose.runtime.*
-import de.hsfl.budgetBinder.presentation.viewmodel.LogoutViewModel
 import de.hsfl.budgetBinder.presentation.Screen
-import de.hsfl.budgetBinder.presentation.UiState
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import org.kodein.di.compose.localDI
-import org.kodein.di.instance
 
 @Composable
 fun UserComponent(screenState: MutableState<Screen>) {
